cluster: Move to ocm package
Since cluster is a wrapper around OCM API functions, we move it under
the ocm package to be consistent with the other API resources.
