+++ cmd/create/cluster/cmd.go
@@ -33,11 +33,11 @@ import (
 	installLogs "github.com/openshift/rosa/cmd/logs/install"
 	"github.com/openshift/rosa/pkg/arguments"
 	"github.com/openshift/rosa/pkg/aws"
-	clusterprovider "github.com/openshift/rosa/pkg/cluster"
 	"github.com/openshift/rosa/pkg/interactive"
 	"github.com/openshift/rosa/pkg/interactive/confirm"
 	"github.com/openshift/rosa/pkg/logging"
 	"github.com/openshift/rosa/pkg/ocm"
+	clusterprovider "github.com/openshift/rosa/pkg/ocm/cluster"
 	"github.com/openshift/rosa/pkg/ocm/machines"
 	"github.com/openshift/rosa/pkg/ocm/regions"
 	"github.com/openshift/rosa/pkg/ocm/versions"
