+++ cmd/logs/uninstall/cmd.go
@@ -28,9 +28,9 @@ import (
 	errors "github.com/zgalor/weberr"
 
 	"github.com/openshift/rosa/pkg/aws"
-	clusterprovider "github.com/openshift/rosa/pkg/cluster"
 	"github.com/openshift/rosa/pkg/logging"
 	"github.com/openshift/rosa/pkg/ocm"
+	clusterprovider "github.com/openshift/rosa/pkg/ocm/cluster"
 	rprtr "github.com/openshift/rosa/pkg/reporter"
 )
